Omstrukturering av serverfiler
