+++ server/src/server/server_axios.ts
@@ -1,0 +1,29 @@
+import express, { Request, Response } from "express";
+import cors from "cors";
+import delete_data from "../database/data_deletion";
+import { table_name_supply_posts } from "../database/connection_pooling";
+
+export const app = express();
+const PORT = 3000;
+
+// Middleware which enables automatic JSON parsing
+app.use(express.json());
+// Enables cross-origin resource sharing
+// i.e. enables the front end (different port) to communicate
+// to the server
+app.use(cors());// TODO: this is not safe in production
+
+app.post("/reset_table", async (req: Request, res: Response) => {
+    try {
+        await delete_data(table_name_supply_posts, {});
+        res.status(200).json();
+    } catch (error: any) {
+		console.error("Failed to reset table", error.message);
+        res.status(500).json();
+    }
+});
+
+// Listens on the port 3000
+app.listen(PORT, () => {
+	console.log(`Server is running on http://localhost:${PORT}`);
+});
